requirements.txt
aws-cdk-lib==2.186.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha >= 2.186.0a0
aws_lambda_powertools
